renames up arrow icon for character update/level-up
