+++ src/app/components/character/CharacterHeaderInfo.tsx
@@ -2,7 +2,7 @@
 
 import Button from "@/app/components/shared/Button";
 import ImageLoadingSkeleton from "@/app/components/shared/ImageLoadingSkeleton";
-import { UpArrowIcon } from "@/app/components/shared/CoiledArrowIcon";
+import { UpArrowIcon } from "@/app/components/shared/UpArrowIcon";
 import Image from "next/image";
 import React, { useState } from "react";
 import { CharacterNameActionsModal } from "./CharacterNameActionsModal";
